requirements.txt
gradio~=5.23.3
torch~=2.6.0
transformers~=4.50.3
pillow~=11.1.0